+++ CHANGELOG.md
@@ -65,3 +65,4 @@
 - Bug #61: Improve `HasIdTest` and `IdProvider` for improved `id` attribute handling and validation (@terabytesoftw)
 - Bug #62: Update assertion message in `ClassProvider` for clarity on `null` class attribute handling (@terabytesoftw)
 - Bug #63: Ensure `id` attribute is set correctly in `HasId` trait for nullable values (@terabytesoftw)
+- Enh #64: Add `HasLang` trait with `lang()` method and tests (@terabytesoftw)